% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.renyi.R
\name{autoplot.renyi}
\alias{autoplot.renyi}
\alias{fortify.renyi}
\alias{autoplot.renyiaccum}
\alias{fortify.renyiaccum}
\title{Autoplot Graphics for vegan renyi and renyiaccum Objects}
\usage{
\method{autoplot}{renyi}(object, fill = "skyblue", alpha = 0.2, ...)

\method{fortify}{renyi}(model, data, ...)

\method{autoplot}{renyiaccum}(
  object,
  ribbon = c("0.95", "minmax", "stdev"),
  fill = "skyblue",
  alpha = 0.5,
  ...
)

\method{fortify}{renyiaccum}(model, data, ...)
}
\arguments{
\item{object}{Result from \pkg{vegan} \code{\link[vegan]{renyi}} or
\code{\link[vegan]{renyiaccum}} functions.}

\item{fill, alpha}{Parameters passed to \code{\link[ggplot2]{geom_ribbon}}}

\item{...}{other arguments passed to methods}

\item{model}{model or other R object to convert to data frame}

\item{data}{original dataset, if needed}

\item{ribbon}{Show ribbon for 0.95 interval, extreme values or for
standard deviation of permutations.}
}
\description{
Alternatives to \pkg{lattice} graphics functions in
\pkg{vegan}.
}
\details{
\code{autoplot} for \code{\link[vegan]{renyi}} displays the
diversites against alpha as dots over a ribbon of data extremes in
the graph and line of data median faceted by sites. \code{autoplot}
for \code{\link[vegan]{renyiaccum}} shows the median of diversity
accumulation against the number of sites over a ribbon of
permutation values faceted by alpha.
}
\examples{
## Rényi
data(BCI)
mod <- renyi(BCI[sample(50, 16),])
autoplot(mod)
## Diversity against the number of sites
mod <- renyiaccum(BCI)
autoplot(mod)
autoplot(mod) + scale_x_log10()

}
